+++ Mailer/app/forms/TemplateFormFactory.php
@@ -84,6 +84,12 @@ class TemplateFormFactory implements IFormFactory
         $form->addText('subject', 'Subject')
             ->setRequired("Field 'Subject' is required.");
 
+        $form->addSelect('click_tracking', 'Click tracking', [
+            null => 'Default mailer settings (depends on your provider configuration)',
+            1 => "Enabled",
+            0 => "Disabled",
+        ]);
+
         $form->addTextArea('mail_body_text', 'Text version')
             ->setAttribute('rows', 3);
 
@@ -130,7 +136,8 @@ class TemplateFormFactory implements IFormFactory
                     $values['mail_body_text'],
                     $values['mail_body_html'],
                     $values['mail_layout_id'],
-                    $values['mail_type_id']
+                    $values['mail_type_id'],
+                    $values['click_tracking']
                 );
                 ($this->onCreate)($row, $buttonSubmitted);
             }