+++ Mailer/migrations/20200914090208_template_click_tracking_flag.php
@@ -1,0 +1,13 @@
+<?php
+
+use Phinx\Migration\AbstractMigration;
+
+class TemplateClickTrackingFlag extends AbstractMigration
+{
+    public function change()
+    {
+        $this->table('mail_templates')
+            ->addColumn('click_tracking', 'boolean', ['null' => true])
+            ->update();
+    }
+}